backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.0.1
python-dotenv==1.0.0
openai==1.3.0
aiofiles==23.2.0
python-magic==0.4.27
requests>=2.28.0
